requirement.txt
# Core Dependencies
pandas==2.2.3
numpy==1.26.4
scikit-surprise==1.1.3
scikit-learn==1.5.2